requirement.txt
torch
torchvision
streamlit
pillow
numpy
pandas
matplotlib